requirements.txt
asgiref==3.5.2
Django==3.2.15
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.30
mysqlclient==2.1.0
Pillow==9.2.0
protobuf==3.20.1
PyMySQL==1.0.2
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
